refactor: move query module into testify feature
