+++ lua/neotest-golang/features/testify/init.lua
@@ -1,5 +1,6 @@
 local M = {}
 
+M.namespace = require("neotest-golang.features.testify.namespace")
 M.lookup = require("neotest-golang.features.testify.lookup")
 M.query = require("neotest-golang.features.testify.query")
 M.tree_modification =
